requirements.txt
pynest-api==0.1.2
asyncpg==0.29.0
injector==0.21.0
fastapi==0.95.1
black==21.12b0
SQLAlchemy==2.0.19
motor==3.2.0
beanie==1.20.0
click==8.1.6
PyYAML==6.0.1
injector==0.21.0
pydantic<2.0.0